Added new Target "InsertUpdateTable.php"
This target is able to import records from a .CSV file into the target table which can be specified on the target section in your strategy.
If a record does not exist in the table, it will be inserted, otherwise it will be just updated. It is possible to specify exception fields in the configuration which are never being updated.
Example usage included in the description of the file.